Factor Initializers out of goal_runner.py.
