+++ src/python/pants/bin/goal_runner.py
@@ -7,238 +7,34 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 import logging
 import sys
-from contextlib import contextmanager
 
-import pkg_resources
 from twitter.common.collections import OrderedSet
 
-from pants.base.build_environment import get_buildroot, get_scm, pants_version
+from pants.base.build_environment import get_scm
 from pants.base.cmd_line_spec_parser import CmdLineSpecParser
-from pants.base.exceptions import BuildConfigurationError
 from pants.base.file_system_project_tree import FileSystemProjectTree
 from pants.base.scm_project_tree import ScmProjectTree
 from pants.base.workunit import WorkUnit, WorkUnitLabel
-from pants.bin.extension_loader import load_plugins_and_backends
-from pants.bin.plugin_resolver import PluginResolver
+from pants.bin.engine_initializer import EngineInitializer
 from pants.bin.repro import Reproducer
 from pants.build_graph.address_lookup_error import AddressLookupError
 from pants.build_graph.build_file_address_mapper import BuildFileAddressMapper
 from pants.build_graph.build_file_parser import BuildFileParser
 from pants.build_graph.mutable_build_graph import MutableBuildGraph
-from pants.engine.exp.engine import LocalSerialEngine
-from pants.engine.exp.fs import create_fs_tasks
-from pants.engine.exp.graph import create_graph_tasks
-from pants.engine.exp.legacy.graph import ExpGraph, create_legacy_graph_tasks
-from pants.engine.exp.legacy.parser import LegacyPythonCallbacksParser, TargetAdaptor
-from pants.engine.exp.mapper import AddressMapper
-from pants.engine.exp.parser import SymbolTable
-from pants.engine.exp.scheduler import LocalScheduler
-from pants.engine.exp.storage import Storage
 from pants.engine.round_engine import RoundEngine
 from pants.goal.context import Context
 from pants.goal.goal import Goal
 from pants.goal.run_tracker import RunTracker
 from pants.help.help_printer import HelpPrinter
 from pants.java.nailgun_executor import NailgunProcessGroup
-from pants.logging.setup import setup_logging
-from pants.option.global_options import GlobalOptionsRegistrar
-from pants.option.options_bootstrapper import OptionsBootstrapper
 from pants.pantsd.subsystem.pants_daemon_launcher import PantsDaemonLauncher
-from pants.reporting.report import Report
 from pants.reporting.reporting import Reporting
 from pants.source.source_root import SourceRootConfig
-from pants.subsystem.subsystem import Subsystem
 from pants.task.task import QuietTaskMixin
 from pants.util.filtering import create_filters, wrap_filters
-from pants.util.memo import memoized_method
 
 
 logger = logging.getLogger(__name__)
-
-
-# N.B. This should be top-level in the module for pickleability - don't nest it.
-class LegacySymbolTable(SymbolTable):
-  """A v1 SymbolTable facade for use with the v2 engine."""
-
-  @classmethod
-  @memoized_method
-  def aliases(cls):
-    """TODO: This is a nasty escape hatch to pass aliases to LegacyPythonCallbacksParser."""
-    _, build_config = OptionsInitializer(OptionsBootstrapper(), init_logging=False).setup()
-    return build_config.registered_aliases()
-
-  @classmethod
-  @memoized_method
-  def table(cls):
-    return {alias: TargetAdaptor for alias in cls.aliases().target_types}
-
-
-class EngineInitializer(object):
-  """Constructs the components necessary to run the v2 engine with v1 BuildGraph compatibility."""
-
-  @staticmethod
-  def parse_commandline_to_spec_roots(options=None, args=None, build_root=None):
-    if not options:
-      options, _ = OptionsInitializer(OptionsBootstrapper(args=args), init_logging=False).setup()
-    cmd_line_spec_parser = CmdLineSpecParser(build_root or get_buildroot())
-    spec_roots = [cmd_line_spec_parser.parse_spec(spec) for spec in options.target_specs]
-    return spec_roots
-
-  @staticmethod
-  def _setup_scheduler(build_root):
-    storage = Storage.create(debug=False)
-    # Ignore any dotfile below build_root except `.` itself
-    project_tree = FileSystemProjectTree(build_root, ['.*'])
-    symbol_table_cls = LegacySymbolTable
-
-    # Register "literal" subjects required for these tasks.
-    # TODO: Replace with `Subsystems`.
-    address_mapper = AddressMapper(symbol_table_cls=symbol_table_cls,
-                                   parser_cls=LegacyPythonCallbacksParser)
-
-    # Create a Scheduler containing graph and filesystem tasks, with no installed goals. The
-    # ExpGraph will explicitly request the products it needs.
-    tasks = (
-      create_legacy_graph_tasks() +
-      create_fs_tasks() +
-      create_graph_tasks(address_mapper, symbol_table_cls)
-    )
-
-    return (
-      LocalScheduler(dict(), tasks, storage, project_tree),
-      storage,
-      symbol_table_cls
-    )
-
-  @classmethod
-  def setup_legacy_graph(cls):
-    build_root = get_buildroot()
-    scheduler, storage, symbol_table_cls = cls._setup_scheduler(build_root)
-    engine = LocalSerialEngine(scheduler, storage)
-    return (scheduler, engine, symbol_table_cls, ExpGraph)
-
-  @classmethod
-  @contextmanager
-  def open_legacy_graph(cls, options=None):
-    spec_roots = cls.parse_commandline_to_spec_roots(options)
-    scheduler, engine, symbol_table_cls, build_graph_cls = cls.setup_legacy_graph()
-
-    engine.start()
-    try:
-      graph = build_graph_cls(scheduler, engine, symbol_table_cls)
-      addresses = tuple(graph.inject_specs_closure(spec_roots))
-      yield graph, addresses, scheduler
-    finally:
-      logger.debug('engine cache stats: {}'.format(engine._cache.get_stats()))
-      engine.close()
-
-
-class OptionsInitializer(object):
-  """Initializes global options and logging."""
-
-  def __init__(self, options_bootstrapper, working_set=None, exiter=sys.exit, init_logging=True):
-    """
-    :param OptionsBootStrapper options_bootstrapper: An options bootstrapper instance.
-    :param pkg_resources.WorkingSet working_set: The working set of the current run as returned by
-                                                 PluginResolver.resolve().
-    :param func exiter: A function that accepts an exit code value and exits (for tests).
-    """
-    self._options_bootstrapper = options_bootstrapper
-    self._working_set = working_set or PluginResolver(self._options_bootstrapper).resolve()
-    self._exiter = exiter
-    self._init_logging = init_logging
-
-  def _setup_logging(self, global_options):
-    """Sets global logging."""
-    # N.B. quiet help says 'Squelches all console output apart from errors'.
-    level = 'ERROR' if global_options.quiet else global_options.level.upper()
-    setup_logging(level, console_stream=sys.stderr, log_dir=global_options.logdir)
-
-  def _register_options(self, subsystems, options):
-    """Registers global options."""
-    # Standalone global options.
-    GlobalOptionsRegistrar.register_options_on_scope(options)
-
-    # Options for subsystems.
-    for subsystem in subsystems:
-      subsystem.register_options_on_scope(options)
-
-    # TODO(benjy): Should Goals or the entire goal-running mechanism be a Subsystem?
-    for goal in Goal.all():
-      # Register task options.
-      goal.register_options(options)
-
-  def _setup_options(self, options_bootstrapper, working_set):
-    bootstrap_options = options_bootstrapper.get_bootstrap_options()
-    global_bootstrap_options = bootstrap_options.for_global_scope()
-
-    if global_bootstrap_options.pants_version != pants_version():
-      raise BuildConfigurationError(
-        'Version mismatch: Requested version was {}, our version is {}.'.format(
-          global_bootstrap_options.pants_version, pants_version()
-        )
-      )
-
-    # Get logging setup prior to loading backends so that they can log as needed.
-    if self._init_logging:
-      self._setup_logging(global_bootstrap_options)
-
-    # Add any extra paths to python path (e.g., for loading extra source backends).
-    for path in global_bootstrap_options.pythonpath:
-      sys.path.append(path)
-      pkg_resources.fixup_namespace_packages(path)
-
-    # Load plugins and backends.
-    plugins = global_bootstrap_options.plugins
-    backend_packages = global_bootstrap_options.backend_packages
-    build_configuration = load_plugins_and_backends(plugins, working_set, backend_packages)
-
-    # Now that plugins and backends are loaded, we can gather the known scopes.
-    known_scope_infos = [GlobalOptionsRegistrar.get_scope_info()]
-
-    # Add scopes for all needed subsystems via a union of all known subsystem sets.
-    subsystems = Subsystem.closure(
-      GoalRunner.subsystems() | Goal.subsystems() | build_configuration.subsystems()
-    )
-    for subsystem in subsystems:
-      known_scope_infos.append(subsystem.get_scope_info())
-
-    # Add scopes for all tasks in all goals.
-    for goal in Goal.all():
-      known_scope_infos.extend(filter(None, goal.known_scope_infos()))
-
-    # Now that we have the known scopes we can get the full options.
-    options = options_bootstrapper.get_full_options(known_scope_infos)
-    self._register_options(subsystems, options)
-
-    # Make the options values available to all subsystems.
-    Subsystem.set_options(options)
-
-    return options, build_configuration
-
-  def setup(self):
-    return self._setup_options(self._options_bootstrapper, self._working_set)
-
-
-class ReportingInitializer(object):
-  """Starts and provides logged info on the RunTracker and Reporting subsystems."""
-
-  def __init__(self, run_tracker=None, reporting=None):
-    self._run_tracker = run_tracker or RunTracker.global_instance()
-    self._reporting = reporting or Reporting.global_instance()
-
-  def setup(self):
-    """Start up the RunTracker and log reporting details."""
-    report = self._reporting.initial_reporting(self._run_tracker)
-    self._run_tracker.start(report)
-
-    url = self._run_tracker.run_info.get_info('report_url')
-    if url:
-      self._run_tracker.log(Report.INFO, 'See a report at: {}'.format(url))
-    else:
-      self._run_tracker.log(Report.INFO, '(To run a reporting server: ./pants server)')
-
-    return self._run_tracker, self._reporting
 
 
 class GoalRunnerFactory(object):
