+++ tests/python/pants_test/bin/test_options_initializer.py
@@ -9,7 +9,7 @@ import pytest
 from pkg_resources import WorkingSet
 
 from pants.base.exceptions import BuildConfigurationError
-from pants.bin.goal_runner import OptionsInitializer
+from pants.bin.options_initializer import OptionsInitializer
 from pants.option.options_bootstrapper import OptionsBootstrapper
